Story 2 - Preventing deletion of meetings by meeting ID + unit test
Added some modifications
Added unit tests

+++ src/main/java/pl/akademiaspecjalistowit/powtorzeniematerialu/meeting/MeetingService.java
@@ -30,6 +30,9 @@ public class MeetingService {
     }
 
     public void deleteExistingMeeting(Meeting meeting) {
+        if (!getAllMeetings().contains(meeting)) {
+            throw new MeetingException("Niestety nie znaleziono żadnego spotkania!");
+        }
         meetingRepository.remove(meeting);
     }
 }
